Added some rate limiting, need to fix get assignments and add authentication

+++ api/courses.js
@@ -1,8 +1,9 @@
 const router = require('express').Router()
 exports.router = router;
 
+const { requireAuthentication } = require('../lib/auth');
 const { validateAgainstSchema } = require('../lib/validation')
-const { courseSchema, getCoursesPage, insertNewCourse, getCourseById, updateCourse, deleteCourse} = require('../models/course')
+const { courseSchema, getCoursesPage, insertNewCourse, getCourseById, updateCourse, deleteCourse, getCourseAssignments} = require('../models/course')
 
 router.get('/', async (req, res) => {
     try {
@@ -26,10 +27,12 @@ router.get('/', async (req, res) => {
 })
 
 //Add a new course to the database
-router.post('/', async function (req, res, next) {
-    //
-    //Validate if user is an admin here
-    //
+router.post('/', requireAuthentication, async function (req, res, next) {
+  /*  if(!req.admin) {
+        res.status(403).send({
+            error: "Unathorized to access the specified resource"
+        })
+    } */
     if(validateAgainstSchema(req.body, courseSchema)) {
         const id = await insertNewCourse(req.body)
         res.status(201).json({
@@ -54,34 +57,52 @@ router.get('/:courseId', async function (req, res, next) {
     }
 })
 
-router.patch('/:courseId', async function (req, res, next) {
-    //
-    // Validate user here
-    //
-    try {
-        const updateSuccessful = await updateCourse(req.params.courseId, req.body);
-        if (updateSuccessful) {
-            res.status(200).json({
-                links: {
-                    course: `/courses/${req.params.courseId}`
-                }
-            })
-        } else {
-            next()
-        }
-    } catch(err) {
+router.patch('/:courseId', requireAuthentication, async function (req, res, next) {
+  /*  if(!req.admin) {
+        res.status(403).send({
+            error: "Unathorized to access the specified resource"
+        })
+    } */
+    try{
+        const courseid = req.params.courseId
+        const updatecourse = req.body
+        await updateCourse(courseid, updatecourse)
+        res.status(200).json({
+            links: {
+                assignment: `/courses/${courseid}`
+            }
+        });
+    }catch(err){
         next()
     }
 })
 
 router.delete('/:courseId', async function (req, res, next) {
-    //
-    // validate user here
-    //
+    if(!req.admin) {
+        res.status(403).send({
+            error: "Unathorized to access the specified resource"
+        })
+        return
+    }
     const deleteSucess = await deleteCourse(req.params.courseId)
     if (deleteSucess) {
         res.status(204).end()
     } else {
         next()
     }
+})
+
+//router.get
+//router.post
+//router.get
+
+router.get('/:courseId/assignments', async function (req, res, next) {
+        const id = req.params.courseId
+        console.log(id)
+        const assignments = await getCourseAssignments(id)
+        if(assignments) {
+            res.status(200).json(assignments);
+        } else {
+            next()
+        }
 })
